// apps/web/src/api/generated/schemas/suppliersGetSupplierInvoicesType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * X-Ear CRM API
 * Merged: manual schemas + auto-generated endpoints
 * OpenAPI spec version: 1.0.0
 */

export type SuppliersGetSupplierInvoicesType =
  (typeof SuppliersGetSupplierInvoicesType)[keyof typeof SuppliersGetSupplierInvoicesType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SuppliersGetSupplierInvoicesType = {
  incoming: "incoming",
  outgoing: "outgoing",
  all: "all",
} as const;