// apps/web/src/api/generated/schemas/suppliersGetSupplierInvoices200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * X-Ear CRM API
 * Merged: manual schemas + auto-generated endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { SuppliersGetSupplierInvoices200Data } from "./suppliersGetSupplierInvoices200Data";

export type SuppliersGetSupplierInvoices200 = {
  success?: boolean;
  data?: SuppliersGetSupplierInvoices200Data;
  requestId?: string;
  timestamp?: string;
};