// apps/web/src/api/generated/schemas/suppliersGetSuggestedSuppliers200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * X-Ear CRM API
 * Merged: manual schemas + auto-generated endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { SuppliersGetSuggestedSuppliers200Data } from "./suppliersGetSuggestedSuppliers200Data";

export type SuppliersGetSuggestedSuppliers200 = {
  success?: boolean;
  data?: SuppliersGetSuggestedSuppliers200Data;
  requestId?: string;
  timestamp?: string;
};